[REFACTOR] #434: Quote 관련 클래스들과 테스트를 하나의 패키지로 묶기
